The allocator was priced before anything was designed: on
x86_64-windows-gnu with zig 0.16.0's mingw CRT,
cost(request) = roundup16(request + 8) — an 8-byte header, measured by
modal pointer stride over 300,000 same-size blocks and corroborated by
working-set delta, agreeing to 0.8 B. A second block derived the same law
independently by a different probe and a different counter. Both were
nearly fooled the same way: _msize echoes the request and reports zero
slack at every size, so on this target no size-reporting API can see the
allocator's real block — only the layout can.

That constant overturns a merged commit's reasoning. 8bd31b25 rejected
this narrowing as worth nothing, assuming a 16-byte header would put 88
and 96 in one bin. Measured, 88 goes to 96 and 96 goes to 112: different
bins, sixteen bytes apart.

Twelve rather than sixteen, because sixteen is a local minimum. RECV 1:1
retains 450,278 strings of capacity 10, where a 24-byte and a 16-byte
header both cost 48 and only 12 reaches 32. Eight buys a further 0.02%
and costs the in-place concat arm its capacity test.

Peak working set falls 12.5% to 16.6% across the four messaging
scenarios on both backends, against in-session A/A floors of 0.003% to
0.317% — 40x the floor at worst, 4,800x at best — with the two work
sizes agreeing within 0.8 pp, which makes it a property of the change
rather than of the operating point. Pass-through from counted bytes to
resident bytes is measured here, not borrowed: ~100%. Allocation traffic
is bit-identical on all four scenarios, so the change alters block sizes
and nothing else, and no CPU claim is made or needed.

On zapo the change is NOT measurable. The three-arm pairing —
base, the same binary again, and the branch — is identical on every
field, and the census bounds live ScrStr bytes below 64 KB there, three
orders of magnitude under the bench. This is a bench result and is
labelled as one.

References are free at 1.0004 per value, duplication is 0.00-0.04% so
interning recovers nothing, and the population is 100% ASCII, so nothing
was paying for UTF-16.

llvm-differential: branch equals base, four pre-existing failures on
both revisions and zero added. The corpus covers the changed offsets
heavily — 290 programs emit the moved ->data GEP and 349 the narrowed
%ScrStr type — which is coverage neither of this change's own two corpus
programs provides.

+++ packages/runtime/src/scr_ws_global.c
@@ -502,7 +502,8 @@ void scr_ws_global_close(ScrWsGlobal *g, bool has_code, double code, const ScrSt
   if (reason != NULL && reason->len > 123) {
     char m[96];
     int n = snprintf(m, sizeof m,
-                     "Reason must be less than 123 bytes; received %zu", reason->len);
+                     "Reason must be less than 123 bytes; received %zu",
+                     (size_t)reason->len);
     wsg_syntax(m, (size_t)n);
     return;
   }
